requirements.txt
aiohttp==3.7.3
aiohttp-jinja2==1.4.2
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
chardet==3.0.4
decorator==4.4.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
motor==2.3.0
multidict==5.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyaml==20.4.0
Pygments==2.7.3
pymongo==3.11.2
PyYAML==5.3.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uuid==1.30
wcwidth==0.2.5
yarl==1.6.3
requests
